pyproject.toml
[project]
name = "workraft"
version = "0.4.7"
description = "A simple worker library in Python"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [{ name = "Artur A. Galstyan", email = "mail@arturgalstyan.dev" }]
dependencies = [
    "beartype",
    "asyncpg",
    "python-dotenv",
    "pydantic",
    "loguru",
    "fire",
    "tenacity",
    "psycopg2>=2.9.9",
    "pytest-asyncio>=0.24.0",
]
[project.optional-dependencies]
dev = ["pre-commit", "pytest", "mkdocs"]

[tool.ruff]
extend-include = ["*.ipynb"]
src = []

[tool.ruff.lint]
fixable = ["I001", "F401"]
ignore = ["E402", "E721", "E731", "E741", "F722", "F821"]
ignore-init-module-imports = true
select = ["E", "F", "I001"]

[tool.ruff.lint.isort]
combine-as-imports = true
extra-standard-library = ["typing_extensions"]
lines-after-imports = 2
order-by-type = false

[tool.pyright]
include = ["workraft", "tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
workraft = "workraft.__main__:main"
